+++ app/src/main/java/com/tapadoo/alerter/demo/JavaDemoActivity.java
@@ -1,12 +1,18 @@
 package com.tapadoo.alerter.demo;
 
+import android.graphics.Color;
 import android.graphics.Typeface;
+import android.net.Uri;
 import android.os.Bundle;
+
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
+
+import android.view.Gravity;
 import android.view.View;
 import android.widget.Toast;
 
+import com.tapadoo.alerter.Alert;
 import com.tapadoo.alerter.Alerter;
 import com.tapadoo.alerter.OnHideAlertListener;
 import com.tapadoo.alerter.OnShowAlertListener;
@@ -38,6 +44,7 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
         findViewById(com.tapadoo.alerter.demo.R.id.btnAlertWithProgress).setOnClickListener(this);
         findViewById(com.tapadoo.alerter.demo.R.id.btnAlertWithCustomFont).setOnClickListener(this);
         findViewById(com.tapadoo.alerter.demo.R.id.btnAlertSwipeToDismissEnabled).setOnClickListener(this);
+        findViewById(com.tapadoo.alerter.demo.R.id.btnAlertSound).setOnClickListener(this);
     }
 
     @Override
@@ -68,8 +75,16 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
             showAlertWithProgress();
         } else if (i == com.tapadoo.alerter.demo.R.id.btnAlertWithCustomFont) {
             showAlertWithCustomFont();
+        } else if (i == com.tapadoo.alerter.demo.R.id.btnAlertWithCustomColor) {
+            showAlertWithCustomColor();
         } else if (i == com.tapadoo.alerter.demo.R.id.btnAlertSwipeToDismissEnabled) {
             showAlertSwipeToDismissEnabled();
+        } else if (i == com.tapadoo.alerter.demo.R.id.btnAlertSound) {
+            showAlertSound();
+        } else if (i == com.tapadoo.alerter.demo.R.id.btnCenterAlert) {
+            showAlertFromCenter();
+        } else if (i == com.tapadoo.alerter.demo.R.id.btnBottomAlert) {
+            showAlertFromBottom();
         } else {
             showAlertDefault();
         }
@@ -95,6 +110,7 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
                 .setText("Alert text...")
                 .setIcon(com.tapadoo.alerter.demo.R.drawable.alerter_ic_mail_outline)
                 .setIconColorFilter(0) // Optional - Removes white tint
+                .setIconSize(R.dimen.custom_icon_size) // Optional - default is 38dp
                 .show();
     }
 
@@ -122,12 +138,12 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
         Alerter.create(JavaDemoActivity.this)
                 .setTitle("Alert Title")
                 .setText("The alert scales to accommodate larger bodies of text. " +
-                         "The alert scales to accommodate larger bodies of text. " +
-                         "The alert scales to accommodate larger bodies of text.")
+                        "The alert scales to accommodate larger bodies of text. " +
+                        "The alert scales to accommodate larger bodies of text.")
                 .show();
     }
 
-    private void showAlertCallbacks(){
+    private void showAlertCallbacks() {
         Alerter.create(JavaDemoActivity.this)
                 .setTitle("Alert Title")
                 .setText("Alert text...")
@@ -175,6 +191,15 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
                 .show();
     }
 
+    private void showAlertWithCustomColor() {
+        Alert alert = Alerter.create(JavaDemoActivity.this)
+                .setTitle("Yellow Alert Title")
+                .setText("Red Alert text...")
+                .show();
+        alert.getTitle().setTextColor(Color.YELLOW);
+        alert.getText().setTextColor(Color.RED);
+    }
+
     private void showAlertSwipeToDismissEnabled() {
         Alerter.create(JavaDemoActivity.this)
                 .setTitle("Alert Title")
@@ -211,4 +236,28 @@ public class JavaDemoActivity extends AppCompatActivity implements View.OnClickL
                 .show();
     }
 
+    private void showAlertSound() {
+        Alerter.create(JavaDemoActivity.this)
+                .setTitle("Alert Title")
+                .setText("Alert text...")
+                .setBackgroundColorRes(R.color.colorAccent)
+                .setSound(Uri.parse("android.resource://" + getPackageName() + "/raw/ringtone"))
+                .show();
+    }
+
+    private void showAlertFromCenter() {
+        Alerter.create(JavaDemoActivity.this)
+                .setTitle(com.tapadoo.alerter.demo.R.string.title_activity_example)
+                .setText("Alert text...")
+                .setLayoutGravity(Gravity.CENTER)
+                .show();
+    }
+
+    private void showAlertFromBottom() {
+        Alerter.create(JavaDemoActivity.this)
+                .setTitle(com.tapadoo.alerter.demo.R.string.title_activity_example)
+                .setText("Alert text...")
+                .setLayoutGravity(Gravity.BOTTOM)
+                .show();
+    }
 }