+++ CHANGELOG.md
@@ -1,6 +1,59 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 7.2.4 - 12/08/2021
+* Added ability to set background elevation
+
+## 7.2.2 - 09/08/2021
+* bugfix notchHeight() cast
+
+## 7.2.1 - 27/07/2021
+* Fixed alert may has extra bottom margin when displayed with Gravity.BOTTOM
+
+## 7.2.0 - 14/05/2021
+* OnHideAlertListener added to hide() and clearCurrent() methods.
+
+## 7.0.1 - 05/02/2021
+* updated navigation_bar height getter.
+
+## 7.0.0 - 27/01/2021
+* Added support in Dialog.
+
+## 6.2.1 - 03/09/2020
+* Updated negative margin to fix overshoot on Samsung S10
+
+## 6.2.0 - 03/09/2020
+* Added the ability to enable\disable click animation
+
+## 6.1.0 - 27/06/2020
+* Readded getters for title & text TextView
+
+## 6.0.0 - 27/06/2020
+* Support custom sound
+
+## 5.1.3 - 06/06/2020
+* Bug Fixes
+
+## 5.1.2 - 16/02/2020
+* Added support for rightIcon.
+
+## 5.1.1 - 28/01/2020
+* Minor refactoring
+
+## 5.1.0 - 28/01/2020
+* Added support for center or bottom gravity
+
+## 5.0.0 - 18/09/2019
+* Added support for setting custom layout.
+* Added set icon size.
+
+## 4.1.1 - 25/09/2019
+* Enabled Sound for the Alert
+
+## 4.1.0 - 18/09/2019
+* Updated Build Tools & Dependencies
+* Enabled Vector Compat Support
+
 ## 4.0.3 - 1/5/2019
 * Added support for setting custom typefaces for buttons
 
@@ -36,7 +89,7 @@ All notable changes to this project will be documented in this file.
 * Updated build tools & support libs
 * Added methods to set icon tint
 
-## 2.0.2 - 11/12/2017 
+## 2.0.2 - 11/12/2017
 * Added fixed for the Alert drawing under the status bar
 
 ## 2.0.1 - 18/09/2017
@@ -47,14 +100,14 @@ All notable changes to this project will be documented in this file.
 * Added isShowing method
 * Added additional icon setting methods
 * Added styling methods
-* added progress bar
+* Added progress bar
 
 ## 1.0.9 - 27/06/2017
 * Added Swipe to Dismiss
 * Added setAlertBackgroundColor to allow the use of color ints
 
 ## 1.0.8 - 09/05/2017
-* Added disable vibration option 
+* Added disable vibration option
 * Updated Alert dismiss method
 * Added method to hide icon
 * Added method to disable outside touch